backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
websockets==12.0
aiofiles==23.2.1
psutil==5.9.6
PyYAML==6.0.1
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3

# Raspberry Pi Camera
picamera2>=0.3.12
libcamera>=1.0.0

# Optional dependencies for advanced features
# rawpy==0.18.1  # For DNG processing
# astropy==5.3.4  # For astronomical calculations